refactor: rename llm_kernel as llm_core
