+++ aios/llm_core/README.md
@@ -1,4 +1,4 @@
-# aios/llm_kernel/
+# aios/llm_core/
 
 This directory contains wrapper classes that abstract the LLMs into a single class called LLMKernel. Currently, the following models are supported:
