+++ aios/llm_core/llm_classes/model_registry.py
